Extract generate_tests, rename generators to dialects
